Use board specific storage API: getBoardStorageApi()
