+++ speeduino/board_stm32_official.h
@@ -65,11 +65,6 @@ static constexpr uint32_t ticksToMicros(COMPARE_TYPE ticks)
 constexpr uint16_t BLOCKING_FACTOR = 121;
 constexpr uint16_t TABLE_BLOCKING_FACTOR = 64;
 
-//Select one for EEPROM,the default is EEPROM emulation on internal flash.
-//#define SRAM_AS_EEPROM /*Use 4K battery backed SRAM, requires a 3V continuous source (like battery) connected to Vbat pin */
-//#define USE_SPI_EEPROM PB0 /*Use M25Qxx SPI flash on BlackF407VE*/
-//#define FRAM_AS_EEPROM /*Use FRAM like FM25xxx, MB85RSxxx or any SPI compatible */
-
 #ifndef word
   #define word(h, l) (((h) << 8) | (l)) //word() function not defined for this platform in the main library
 #endif  
@@ -184,25 +179,8 @@ extern STM32RTC& rtc;
 ***********************************************************************************************************
 * EEPROM emulation
 */
-#if defined(SRAM_AS_EEPROM)
-  #define EEPROM_LIB_H "src/BackupSram/BackupSramAsEEPROM.h"
-  class BackupSramAsEEPROM;
-  using EEPROM_t = BackupSramAsEEPROM;    
-#elif defined(USE_SPI_EEPROM)
-  #define EEPROM_LIB_H "src/SPIAsEEPROM/SPIAsEEPROM.h"
-  class SPI_EEPROM_Class;
-  using EEPROM_t = SPI_EEPROM_Class;    
-#elif defined(FRAM_AS_EEPROM) //https://github.com/VitorBoss/FRAM
-  #define EEPROM_LIB_H "src/FRAM/Fram.h"
-  class FramClass;
-  using EEPROM_t = FramClass;    
-#else //default case, internal flash as EEPROM
-  #define EEPROM_LIB_H "src/SPIAsEEPROM/SPIAsEEPROM.h"
-  class InternalSTM32F4_EEPROM_Class;
-  using EEPROM_t = InternalSTM32F4_EEPROM_Class;    
-  #if defined(STM32F401xC)
-    #define SMALL_FLASH_MODE
-  #endif
+#if defined(STM32F401xC)
+  #define SMALL_FLASH_MODE
 #endif
 
 #define RTC_LIB_H "STM32RTC.h"
