+++ speeduino/storage_api.cpp
@@ -1,48 +1,8 @@
 #include "storage_api.h"
 #include "board_definition.h"
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wunused-variable"
-#include EEPROM_LIB_H //This is defined in the board .h files
-#pragma GCC diagnostic pop
 
-
-#if defined(CORE_AVR)
-#pragma GCC push_options
 // This minimizes RAM usage at no performance cost
 #pragma GCC optimize ("Os") 
-#endif
-
-// LCOV_EXCL_START
-// Exclude simple wrappers from code coverage
-namespace EEPROMApi {
-
-  static inline byte read(uint16_t address)
-  {
-    return getEEPROM().read(address);
-  }
-  static inline void write(uint16_t address, byte val)
-  {
-    getEEPROM().write(address, val);
-  }
-  static inline uint16_t length(void)
-  {
-    return getEEPROM().length();
-  }
-}
-// LCOV_EXCL_STOP
-
-// LCOV_EXCL_START
-// Exclude simple wrappers from code coverage
-storage_api_t getEEPROMStorageApi(void)
-{
-  return storage_api_t {
-    .read = EEPROMApi::read,
-    .write = EEPROMApi::write,
-    .length = EEPROMApi::length,
-    .getMaxWriteBlockSize = ::getEepromWriteBlockSize,
-  };  
-}
-// LCOV_EXCL_STOP
 
 bool update(const storage_api_t &api, uint16_t address, byte value) {
   if (api.read(address)!=value) {
@@ -108,7 +68,3 @@ __attribute__((noinline)) void fillBlock(const storage_api_t &api, uint16_t addr
     }
   }
 }
-
-#if defined(CORE_AVR)
-#pragma GCC pop_options
-#endif
